+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Share My Info moved to Profile Info page [#5201](https://github.com/rokwire/illinois-app/issues/5201).
 - Updated privacy icons in profile edit page [#5200](https://github.com/rokwire/illinois-app/issues/5200).
 - Updated Food Filters icon [#5206](https://github.com/rokwire/illinois-app/issues/5206).
+### Added
+- Assistant Events Prompt [#5220](https://github.com/rokwire/illinois-app/issues/5220).
 
 ## [7.1.15] - 2025-06-11
 ### Added